% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beard}
\alias{beard}
\title{Beards description}
\format{
A data frame with 494 rows and 3 columns : beard, description (in french), description (in english)
}
\usage{
data(beard)
}
\description{
This dataset deals with the description of 8 beards. Each beard is described by each individual.
}
\keyword{datasets}
