% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensoTable.R
\name{exportxlsx}
\alias{exportxlsx}
\title{Title}
\usage{
exportxlsx(table, filename, path)
}
\arguments{
\item{table}{flextable}

\item{filename}{String}

\item{path}{String}
}
\value{
Returns an .xlsx file based on a flextable
}
\description{
Title
}
