% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensochoc}
\alias{sensochoc}
\title{Sensory data for 6 chocolates}
\format{
A data frame with 348 rows and 19 columns: 5 qualitative variables (Panelist, Session, Form, Rank, Product) and 14 sensory descriptors.
}
\source{
Applied mathematics department, AGROCAMPUS OUEST
}
\usage{
data(sensochoc)
}
\description{
The data used here refer to the sensory description of six varieties of chocolates sold in France: each chocolate was evaluated twice by 29 panelists according to 14 sensory descriptors.
}
\keyword{datasets}
