% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensoTable.R
\name{sensoTable}
\alias{sensoTable}
\title{Title}
\usage{
sensoTable(
  res.decat,
  level.upper = Inf,
  level.upper2 = 0.05,
  col.lower = "#ff7979",
  col.lower2 = "#eb4d4b",
  col.upper = "#7ed6df",
  col.upper2 = "#22a6b3"
)
}
\arguments{
\item{res.decat}{Result of a SensoMineR::decat}

\item{level.upper}{numeric}

\item{level.upper2}{numeric}

\item{col.lower}{string}

\item{col.lower2}{string}

\item{col.upper}{string}

\item{col.upper2}{string}
}
\value{
sensoTable() returns a flextable
}
\description{
Title
}
