% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yes_decat.R
\name{Yes_decat}
\alias{Yes_decat}
\title{PowerPoint reporting of a description of categories}
\usage{
Yes_decat(
  res,
  yes_study_name,
  yes_temp,
  x1 = 1,
  x2 = 2,
  size_tab = 10,
  col.neg = "red",
  col.pos = "blue"
)
}
\arguments{
\item{res}{decat result}

\item{yes_study_name}{title on the first slide}

\item{yes_temp}{path to the template}

\item{x1}{dimension to plot on the x-axis}

\item{x2}{dimension to plot on the y-axis}

\item{size_tab}{maximum number of row of a table on a slide}

\item{col.neg}{color for negative value}

\item{col.pos}{color for positive value}
}
\description{
PowerPoint reporting of a description of categories
}
\examples{
\dontrun{
data("sensochoc")
res.decat <-SensoMineR::decat(sensochoc, formul="~Product+Panelist", firstvar = 5, graph = FALSE)
Yes_decat(res.decat, yes_study_name="Quantitative description of products",yes_temp="path/to/the/template/YesSiR_template.pptx")
}
}
