% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yes_MCA.R
\name{Yes_MCA}
\alias{Yes_MCA}
\title{PowerPoint reporting of a MCA}
\usage{
Yes_MCA(
  res,
  yes_study_name = "MCA + HCPC",
  path = getwd(),
  file.name = "MCA_results.pptx",
  x1 = 1,
  x2 = 2,
  thres_x1 = 2,
  thres_x2 = 2,
  proba = 0.05,
  size_tab = 10
)
}
\arguments{
\item{res}{a MCA object}

\item{yes_study_name}{title displayed on the first slide}

\item{path}{path to the directory which will store the PowerPoint file}

\item{file.name}{name of the PowerPoint file}

\item{x1}{component to plot on the x-axis}

\item{x2}{component to plot on the y-axis}

\item{thres_x1}{the threshold for the x-axis over which points are ploted translucent}

\item{thres_x2}{the threshold for the y-axis over which points are ploted translucent}

\item{proba}{the significance threshold considered to characterized the category (by default 0.05)}

\item{size_tab}{maximum number of rows of a table per slide}
}
\value{
Returns a .pptx file
}
\description{
PowerPoint reporting of a MCA
}
\examples{
\dontrun{
data(tea)
res.mca <- FactoMineR::MCA(tea[,-19],quali.sup=19:35,graph=FALSE)
# Create the PowerPoint in the current working directory
Yes_MCA(res.mca)
}
}
