% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yes_PCA.R
\name{Yes_PCA}
\alias{Yes_PCA}
\title{PowerPont reporting of a PCA}
\usage{
Yes_PCA(
  res,
  yes_study_name = "PCA + HCPC",
  path = getwd(),
  file.name = "PCA_results.pptx",
  x1 = 1,
  x2 = 2,
  proba = 0.05,
  size_tab = 10
)
}
\arguments{
\item{res}{a PCA object}

\item{yes_study_name}{title displayed on the first slide}

\item{path}{PowerPoint file to be created}

\item{file.name}{name of the PowerPoint file}

\item{x1}{component to plot on the x-axis}

\item{x2}{component to plot on the y-axis}

\item{proba}{the significance threshold considered to characterized the category (by default 0.05)}

\item{size_tab}{maximum number of rows of a table per slide}
}
\value{
Returns a .pptx file
}
\description{
PowerPont reporting of a PCA
}
\examples{
\dontrun{
library(FactoMineR)
data(decathlon)
res.pca <- FactoMineR::PCA(decathlon, quanti.sup = 11:12, quali.sup=13)
# Create the PowerPoint in the current working directory
Yes_PCA(res.pca)
}
}
