% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yes_textual.R
\name{Yes_textual}
\alias{Yes_textual}
\title{PowerPoint reporting of a textual analysis}
\usage{
Yes_textual(
  res,
  yes_study_name = "Result of a textual analysis",
  yes_temp,
  path,
  size_tab = 10,
  proba = 0.05,
  col.neg = "red",
  col.pos = "blue"
)
}
\arguments{
\item{res}{FactoMineR::textual result}

\item{yes_study_name}{title displayed on the first slide}

\item{yes_temp}{path to the template}

\item{path}{PowerPoint file to be created}

\item{size_tab}{maximum number of rows of a table per slide}

\item{proba}{the significance threshold considered to characterized the category (by default 0.05)}

\item{col.neg}{color for negative values (on the wordclouds)}

\item{col.pos}{color for positive values (on the wordclouds)}
}
\value{
Returns a .pptx file
}
\description{
PowerPoint reporting of a textual analysis
}
\examples{
\dontrun{
data(beard)
res.text <- textual(beard,contingence.by = 1,num.text = 3)
Yes_textual(res = res.text, yes_study_name = "Textual analysis", yes_temp = "path/to/the/template/YesSiR_template.pptx", path = "path/to/the/powerpoint_file.pptx")
}
}
