% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yes_decat.R
\name{Yes_decat}
\alias{Yes_decat}
\title{PowerPoint reporting of a description of categories}
\usage{
Yes_decat(
  res,
  yes_study_name = "Quantitative description of products",
  path = getwd(),
  file.name = "decat_results.pptx",
  x1 = 1,
  x2 = 2,
  size_tab = 10,
  col.neg = "red",
  col.pos = "blue"
)
}
\arguments{
\item{res}{decat result}

\item{yes_study_name}{title displayed on the first slide}

\item{path}{PowerPoint file to be created}

\item{file.name}{name of the PowerPoint file}

\item{x1}{dimension to plot on the x-axis}

\item{x2}{dimension to plot on the y-axis}

\item{size_tab}{maximum number of rows of a table per slide}

\item{col.neg}{color for negative values (on the wordclouds)}

\item{col.pos}{color for positive values (on the wordclouds)}
}
\description{
PowerPoint reporting of a description of categories
}
\examples{
\dontrun{
data("sensochoc")
res.decat <- SensoMineR::decat(sensochoc, formul="~Product+Panelist", firstvar = 5, graph = FALSE)
# Create the PowerPoint in the current working directory
Yes_decat(res.decat)
}
}
