% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_MCA.R
\name{Yes_MCA}
\alias{Yes_MCA}
\title{Title}
\usage{
Yes_MCA(
  res,
  yes_study_name,
  yes_temp,
  x1 = 1,
  x2 = 2,
  thres_x1 = 2,
  thres_x2 = 2,
  proba = 0.05,
  size_tab = 10
)
}
\arguments{
\item{res}{a MCA object}

\item{yes_study_name}{A string of the title on the first slide}

\item{yes_temp}{Path to the template}

\item{x1}{numeric}

\item{x2}{numeric}

\item{thres_x1}{numeric}

\item{thres_x2}{numeric}

\item{proba}{numeric}

\item{size_tab}{numeric}
}
\value{
yes_MCA returns a .pptx file containing all of the results of a MCA
}
\description{
Title
}
