% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_MCA.R
\name{Yes_MCA}
\alias{Yes_MCA}
\title{PowerPoint reporting of a MCA}
\usage{
Yes_MCA(
  res,
  yes_study_name,
  yes_temp,
  x1 = 1,
  x2 = 2,
  thres_x1 = 2,
  thres_x2 = 2,
  proba = 0.05,
  size_tab = 10
)
}
\arguments{
\item{res}{a MCA object}

\item{yes_study_name}{the title on the first slide}

\item{yes_temp}{path to the template}

\item{x1}{integer specifying the component to plot on the x-axis}

\item{x2}{integer specifying the component to plot on the y-axis}

\item{thres_x1}{the threshold for the x-axis over which points are ploted translucent}

\item{thres_x2}{the threshold for the y-axis over which points are ploted translucent}

\item{proba}{the significance threshold considered to characterized the category (used for the catdes function)}

\item{size_tab}{the maximum number of rows a table should have on a single slide}
}
\value{
Returns a .pptx file containing all of the results of a MCA
}
\description{
PowerPoint reporting of a MCA
}
\examples{
data(tea)
res.mca <- FactoMineR::MCA(tea[,-19],quali.sup=19:35,graph=FALSE)
Yes_MCA(res.mca,yes_study_name="MCA + HCPC",yes_temp="C:/SSD/Stage AO/template/YesSir_template.pptx",proba = 0.05,x1=1,x2=2)
}
