% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yes_MCA.R
\name{Yes_MCA}
\alias{Yes_MCA}
\title{Reporting with officer}
\usage{
Yes_MCA(
  res,
  yes_study_name,
  yes_temp,
  x1 = 1,
  x2 = 2,
  thres_x1 = 2,
  thres_x2 = 2,
  proba = 0.05,
  size_tab = 10
)
}
\arguments{
\item{res}{Objet of class MCA}

\item{yes_study_name}{Title}

\item{yes_temp}{A .pptx file}

\item{x1}{Dimension 1}

\item{x2}{Dimension 2}

\item{thres_x1}{Numeric}

\item{thres_x2}{Numeric}

\item{proba}{Numeric}

\item{size_tab}{Integer}
}
\value{
A .pptx file
}
\description{
Reporting with officer based on outputs generated by the FactoMineR package
}
\examples{
#Yes_MCA(res.mca,yes_study_name="MCA + HCPC",yes_temp="YesSiR_template.pptx",proba = 0.05,x1=1,x2=2)
}
